+++ SignalR-Swift/Client/Protocols/ClientTransportProtocol.swift
@@ -13,6 +13,6 @@ protocol ClientTransportProtocol {
     func start(connection: ConnectionProtocol, connectionData: String, completionHandler: ((_ response: AnyObject, _ error: Error?) -> ()))
     func send(connection: ConnectionProtocol, data: String, connectionData: String, completionHandler: ((_ response: AnyObject, _ error: Error?) -> ()))
 
-    func abort(connection: ConnectionProtocol, timeout: Int, connectionData: String)
+    func abort(connection: ConnectionProtocol, timeout: Double, connectionData: String)
     func lostConnection(connection: ConnectionProtocol)
 }
